// src/components/Model/Restaurant.ts
export interface Restaurant {
  id: number;
  restaurantName: string;
  restaurantAddress1: string;
  restaurantAddress2: string;
  restaurantAddress3: string;
  restaurantAddress4: string;
  restaurantCity: string;
  restaurantState: string;
  restaurantPostalCode: string;
  fssaifile: string;
  gstfile: string;
  aadharPANFile:string;
  bankPANFile:string;
  l1Approve:string
  l2Approve:string;
  isActive:string;
  email: string;
  cuisine:string;
  image: string;
  time: string;
  rating: string;
}